# Optimise C# in a ridiculous way