@RReverser 2024-10-31
> How did you even find this?

It started from digging into why iterators specifically are so slow, making some optimisations, being weirded out by not noticing any improvement, and eventually ended with a "scientific guess" that if it all takes roughly constant amount of time, GC must be involved somehow. And voila, indeed it was.

> And to be clear, is it removing the finalizer, or adding the `sealed` which has this effect?

It's removing the finalizer.

`sealed` was added after the fact, because after removing the finalizer I also removed `GC.SuppressFinalize(this);` - which is no longer necessary - and then linter would warn about subclasses potentially adding their own finalizers. So I made the class `sealed` just to tell the linter it's okay and there will be no subclasses.